python_packages/pantrysoft_authenticator/pyproject.toml
[tool.poetry]
name = "pantrysoft_authenticator"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
selenium = "^4.26.1"
pytest-playwright = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
